pyproject.toml
[project]
name = "sip-house-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiofiles>=25.1.0",
    "aiomysql>=0.3.2",
    "alembic>=1.17.2",
    "bcrypt>=5.0.0",
    "cryptography>=46.0.3",
    "fastapi[standard,standart]>=0.128.0",
    "gunicorn>=23.0.0",
    "pillow>=12.1.0",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "pymysql>=1.1.2",
    "pyright>=1.1.407",
    "python-magic>=0.4.27",
    "python-multipart>=0.0.21",
    "sqlalchemy>=2.0.45",
    "typing-extensions>=4.15.0",
    "uvicorn>=0.40.0",
]
